spec: Remove "importAs" field. Add well known fields. Add package vendor comment.
It is still unclear at this time if the "importAs" field will be useful.
The field can always be added in later if tools use it.

There are two concreate use cases for origin. Note them to help
start them off with consistant names.

Specify how the package vendor comment should be used and what it is.
When a rewritten package doesn't live in the local tree checks still
need to be made against it. Allow for this case.

closes #11
closes #6